Extract function for re-drawing an element if its value changed
