+++ firmware/src/gui/pages/base.hh
@@ -97,18 +97,5 @@ struct PageBase {
 
 	virtual void update() {
 	}
-
-	//TODO: add helpers? or add to separate class/namespace?
-	// std::string_view read_slug() {
-	// 	auto module_id = PageList::get_selected_module_id();
-	// 	auto patch_id = PageList::get_selected_patch_id();
-	// 	const PatchData &patch = patch_list.get_patch(patch_id);
-	// 	if (patch.patch_name.length() == 0)
-	// 		return "";
-	// 	if (module_id >= patch.module_slugs.size())
-	// 		return "";
-
-	// 	return patch.module_slugs[module_id];
-	// }
 };
 } // namespace MetaModule
