+++ firmware/src/gui/pages/module_view.hh
@@ -174,15 +174,11 @@ struct ModuleViewPage : PageBase {
 
 		if (is_patch_playing) {
 			for (auto &drawn_el : drawn_elements) {
-				std::visit(
-					[this, gui_el = drawn_el.gui_element](auto &el) {
-						// bool did_update = update_element(el, params, patch, gui_el);
-						// if (did_update) {
-						// 	if (settings.map_ring_flash_active)
-						// 		MapRingDisplay::flash_once(gui_el.map_ring, settings.map_ring_style, true);
-						// }
-					},
-					drawn_el.element);
+				auto was_redrawn = std::visit(UpdateElement{params, patch, drawn_el.gui_element}, drawn_el.element);
+
+				if (was_redrawn && settings.map_ring_flash_active) {
+					MapRingDisplay::flash_once(drawn_el.gui_element.map_ring, settings.map_ring_style, true);
+				}
 			}
 		}
 	}
